updating to use the GroupRetriever
