--- source/Octopus.Server.Extensibility.Authentication.DirectoryServices/DirectoryServices/IDirectoryServicesUserSecurityGroupExpiryChecker.cs
@@ -1,10 +1,0 @@
-using Octopus.Data.Model.User;
-
-namespace Octopus.Server.Extensibility.Authentication.DirectoryServices.DirectoryServices
-{
-    public interface IDirectoryServicesUserSecurityGroupExpiryChecker
-    {
-        bool ShouldFetchExternalGroups(IUser user);
-        bool ShouldFetchExternalGroupsInBackground(IUser user);
-    }
-}
