+++ source/Octopus.Server.Extensibility.Authentication.DirectoryServices/packages.config
@@ -10,7 +10,7 @@
   <package id="Octopus.Data" version="1.0.19" targetFramework="net451" />
   <package id="Octopus.Diagnostics" version="1.0.11" targetFramework="net451" />
   <package id="Octopus.Server.Extensibility" version="2.0.0" targetFramework="net451" />
-  <package id="Octopus.Server.Extensibility.Authentication" version="3.0.0" targetFramework="net451" />
+  <package id="Octopus.Server.Extensibility.Authentication" version="3.0.1-enh-extsecgrpsch0001" targetFramework="net451" />
   <package id="Octopus.Time" version="1.0.8" targetFramework="net451" />
   <package id="System.Collections" version="4.0.11" targetFramework="net451" />
   <package id="System.Runtime" version="4.1.0" targetFramework="net451" />
